refactor: organize scripts
